+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 1.20.5
 `CHG` folding: supports `-- #region`
+`CHG` completion: `auto require` can be disabled
 `FIX` [#522](https://github.com/sumneko/lua-language-server/issues/522)
 `FIX` [#523](https://github.com/sumneko/lua-language-server/issues/523)